EU5 1.0.10 save format support
there are new reduced-byte f64 representations as well as new 24-bit lookup.
Got ids from https://github.com/rakaly/jomini/blob/master/src/binary/lexer.rs

+++ pdx_parser_core/src/strings_resolver.rs
@@ -37,8 +37,8 @@ impl StringsResolver {
         return Ok(StringsResolver { inner });
     }
 
-    pub fn get<'a>(&'a self, id: u16) -> Option<&'a str> {
-        return self.inner.borrow_strings().get(id as usize).copied();
+    pub fn get<'a>(&'a self, id: usize) -> Option<&'a str> {
+        return self.inner.borrow_strings().get(id).copied();
     }
 
     pub fn len(&self) -> usize {
@@ -49,13 +49,13 @@ impl StringsResolver {
         return self.inner.borrow_strings().is_empty();
     }
 
-    pub fn iter(&self) -> impl Iterator<Item = (u16, &str)> {
+    pub fn iter(&self) -> impl Iterator<Item = (usize, &str)> {
         return self
             .inner
             .borrow_strings()
             .iter()
             .enumerate()
-            .map(|(i, &s)| (i as u16, s));
+            .map(|(i, &s)| (i, s));
     }
 }
 impl Default for StringsResolver {
